\name{exponential.sample}
\alias{exponential.sample}
\docType{data}
\title{Random Sample from an Exponential Distribution}
\description{
  This data set is a random sample of observations taken from
  an exponential distribution with a rate of 1.
}
\usage{exponential.sample}
\format{A numeric vector containing 100000 observations.}
\examples{hist(exponential.sample)}
\keyword{datasets}
