\name{mixture.sample}
\alias{mixture.sample}
\docType{data}
\title{Random Sample from a Uniform/Normal Mixture}
\description{
  This data set is a random sample of observations taken from
  an even mixture of a uniform and a normal distribution.
}
\usage{mixture.sample}
\format{A numeric vector containing 100000 observations.}
\details{
  The normal distribution has a mean of 0.5 and a variance
  of 0.001, whereas the uniform distribution is defined on
  the interval [0,1].
}
\examples{hist(mixture.sample)}
\keyword{datasets}
