\name{normal.sample}
\alias{normal.sample}
\docType{data}
\title{Random Sample from a Normal Distribution}
\description{
  This data set is a random sample of observations taken from
  a normal distribution with a mean of 0 and a standard
  deviation of 1.
}
\usage{normal.sample}
\format{A numeric vector containing 100000 observations.}
\examples{hist(normal.sample)}
\keyword{datasets}
