\name{uniform.sample}
\alias{uniform.sample}
\docType{data}
\title{Random Sample from a Uniform Distribution}
\description{
  This data set is a random sample of observations taken from
  a uniform distribution on the interval [0,1].
}
\usage{uniform.sample}
\format{A numeric vector containing 100000 observations.}

\examples{hist(uniform.sample)}
\keyword{datasets}
